# Conflicts:
#	core/src/main/java/com/seibel/distanthorizons/core/api/external/methods/config/client/DhApiDebuggingConfig.java
#	core/src/main/java/com/seibel/distanthorizons/core/config/Config.java
#	core/src/main/java/com/seibel/distanthorizons/core/dataObjects/fullData/sources/CompleteFullDataSource.java
#	core/src/main/java/com/seibel/distanthorizons/core/dataObjects/fullData/sources/HighDetailIncompleteFullDataSource.java
#	core/src/main/java/com/seibel/distanthorizons/core/file/fullDatafile/FullDataFileHandler.java
#	core/src/main/java/com/seibel/distanthorizons/core/file/renderfile/RenderDataMetaFile.java
#	core/src/main/java/com/seibel/distanthorizons/core/generation/WorldGenerationQueue.java
#	core/src/main/java/com/seibel/distanthorizons/core/render/LodRenderSection.java
#	core/src/main/java/com/seibel/distanthorizons/core/render/renderer/DebugRenderer.java
#	core/src/main/java/com/seibel/distanthorizons/core/render/renderer/LodRenderer.java

+++ core/src/main/java/com/seibel/distanthorizons/core/util/FileUtil.java
@@ -49,17 +49,24 @@ public class FileUtil
 		}
 		
 		
-		if (file.renameTo(corruptedFile))
+		if (file.exists())
 		{
-			LOGGER.error("Renamed corrupted file to [" + corruptedFileName + "].");
+			if (file.renameTo(corruptedFile))
+			{
+				LOGGER.error("Renamed corrupted file to [" + corruptedFileName + "].");
+			}
+			else
+			{
+				LOGGER.error("Failed to rename corrupted file to [" + corruptedFileName + "]. Attempting to delete file...");
+				if (!file.delete())
+				{
+					LOGGER.error("Unable to delete corrupted file [" + corruptedFileName + "].");
+				}
+			}
 		}
 		else
 		{
-			LOGGER.error("Failed to rename corrupted file to [" + corruptedFileName + "]. Attempting to delete file...");
-			if (!file.delete())
-			{
-				LOGGER.error("Unable to delete corrupted file [" + corruptedFileName + "].");
-			}
+			LOGGER.error("Corrupted file [" + file + "] doesn't exist.");
 		}
 		
 		return corruptedFile;
